+++ common/src/main/java/com/stevekung/fishofthieves/FishOfThieves.java
@@ -11,6 +11,7 @@ import com.stevekung.fishofthieves.registry.*;
 import com.stevekung.fishofthieves.utils.FOTPlatform;
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.GsonConfigSerializer;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
@@ -64,6 +65,11 @@ public class FishOfThieves
         {
             FishPlaqueRegistry.registerInteractionItem(Items.BUCKET, "alexsmobs:cosmic_cod");
             FishPlaqueRegistry.registerInteractionItem(Items.LAVA_BUCKET, "alexsmobs:stradpole");
+        }
+        // Alex's Caves compatibility
+        if (FOTPlatform.isModLoaded("alexscaves"))
+        {
+            FishPlaqueRegistry.registerInteractionItem(BuiltInRegistries.ITEM.get(new ResourceLocation("alexscaves", "acid_bucket")), "alexscaves:radgill");
         }
 
         var bucket = DispenserBlock.DISPENSER_REGISTRY.get(Items.WATER_BUCKET);